directory tidy
