--- java/pod/node/const.go
@@ -1,25 +1,0 @@
-/**
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package node
-
-const (
-	VIRTUAL_NODE_NAME_PREFIX = "virtual-kubenode-"
-
-	LABEL_VIRTUAL_KUBE_NODE     = "label.io.koupleless/virtual-kubenode"
-	LABEL_VIRTUAL_KUBE_NODE_CPU = "todo"
-	LABEL_VIRTUAL_KUBE_NODE_MEM = "todo"
-
-	TAINT_KUBE_NODE_READY = "taint.io.koupleless/ready"
-)
