+++ test/module_daemonset_suite_test.go
@@ -2,7 +2,7 @@ package test
 
 import (
 	"context"
-	"github.com/koupleless/virtual-kubelet/java/pod/node"
+	"github.com/koupleless/virtual-kubelet/provider/javaBase/node_provider"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
@@ -49,7 +49,7 @@ var _ = Describe("Module DaemonSet", func() {
 		mockBase = NewBaseMock(nodeId, "base", "1.1.1", baseMqttClient)
 		go mockBase.Run()
 		Eventually(func() bool {
-			_, err := k8sClient.CoreV1().Nodes().Get(ctx, node.VIRTUAL_NODE_NAME_PREFIX+nodeId, metav1.GetOptions{})
+			_, err := k8sClient.CoreV1().Nodes().Get(ctx, node_provider.VIRTUAL_NODE_NAME_PREFIX+nodeId, metav1.GetOptions{})
 			return !errors.IsNotFound(err)
 		}, timeout, interval).Should(BeTrue())
 	})
@@ -84,7 +84,7 @@ var _ = Describe("Module DaemonSet", func() {
 			mockScaleBase = NewBaseMock(scaleNodeId, "base", "1.1.1", baseMqttClient)
 			go mockScaleBase.Run()
 			Eventually(func() bool {
-				_, err := k8sClient.CoreV1().Nodes().Get(ctx, node.VIRTUAL_NODE_NAME_PREFIX+scaleNodeId, metav1.GetOptions{})
+				_, err := k8sClient.CoreV1().Nodes().Get(ctx, node_provider.VIRTUAL_NODE_NAME_PREFIX+scaleNodeId, metav1.GetOptions{})
 				return !errors.IsNotFound(err)
 			}, timeout, interval).Should(BeTrue())
 		})
@@ -162,7 +162,7 @@ var _ = Describe("Module DaemonSet", func() {
 				mockScaleBase.Exit()
 			}
 			Eventually(func() bool {
-				_, err := k8sClient.CoreV1().Nodes().Get(ctx, node.VIRTUAL_NODE_NAME_PREFIX+scaleNodeId, metav1.GetOptions{})
+				_, err := k8sClient.CoreV1().Nodes().Get(ctx, node_provider.VIRTUAL_NODE_NAME_PREFIX+scaleNodeId, metav1.GetOptions{})
 				return errors.IsNotFound(err)
 			}, timeout, interval).Should(BeTrue())
 		})
@@ -214,7 +214,7 @@ var _ = Describe("Module DaemonSet", func() {
 			mockBase.Exit()
 		}
 		Eventually(func() bool {
-			_, err := k8sClient.CoreV1().Nodes().Get(ctx, node.VIRTUAL_NODE_NAME_PREFIX+nodeId, metav1.GetOptions{})
+			_, err := k8sClient.CoreV1().Nodes().Get(ctx, node_provider.VIRTUAL_NODE_NAME_PREFIX+nodeId, metav1.GetOptions{})
 			return errors.IsNotFound(err)
 		}, timeout, interval).Should(BeTrue())
 	})
